# internal/api/v1/docs/swagger.yaml
definitions:
...
  CertInfo:
    properties:
      dns_names:
        items:
          type: string
        type: array
      email_addresses:
        items:
          type: string
        type: array
      issuer:
        type: string
      not_after:
        type: integer
      not_before:
        type: integer
      subject:
        type: string
    type: object
...
  ErrorResponse:
    properties:
      error:
        type: string
        x-nullable: true
      message:
        type: string
    type: object
...
  NewAgentRequest:
    properties:
      host:
        type: string
      name:
        type: string
      nightly:
        type: boolean
      port:
        maximum: 65535
        minimum: 1
        type: integer
      type:
        enum:
        - docker
        - system
        type: string
    required:
    - host
    - name
    - port
    - type
    type: object
  NewAgentResponse:
    properties:
      ca:
        $ref: '#/definitions/PEMPairResponse'
      client:
        $ref: '#/definitions/PEMPairResponse'
      compose:
        type: string
    type: object
  PEMPairResponse:
    properties:
      cert:
        format: base64
        type: string
      key:
        format: base64
        type: string
    type: object
...
  SuccessResponse:
    properties:
      details:
        additionalProperties: {}
        type: object
        x-nullable: true
      message:
        type: string
    type: object
...
  VerifyNewAgentRequest:
    properties:
      ca:
        $ref: '#/definitions/PEMPairResponse'
      client:
        $ref: '#/definitions/PEMPairResponse'
      host:
        type: string
    type: object
...
  homepage.FetchResult:
    properties:
      errMsg:
        type: string
      icon:
        items:
          type: integer
        type: array
      statusCode:
        type: integer
    type: object
...
  homepage.IconMetaSearch:
    properties:
      dark:
        type: boolean
      light:
        type: boolean
      png:
        type: boolean
      ref:
        type: string
      source:
        $ref: '#/definitions/homepage.IconSource'
      svg:
        type: boolean
      webP:
        type: boolean
    type: object
  homepage.IconSource:
    enum:
    - https://
    - '@target'
    - '@walkxcode'
    - '@selfhst'
    type: string
    x-enum-varnames:
    - IconSourceAbsolute
    - IconSourceRelative
    - IconSourceWalkXCode
    - IconSourceSelfhSt
...
info:
  contact: {}
paths:
  /agent/create:
    post:
      consumes:
      - application/json
      description: |-
        Create a new agent and return the docker compose file, encrypted CA and client PEMs
        The returned PEMs are encrypted with a random key and will be used for verification when adding a new agent
      parameters:
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/NewAgentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/NewAgentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create a new agent
      tags:
      - agent
      x-id: create
...
  /agent/verify:
    post:
      consumes:
      - application/json
      description: Verify a new agent and return the number of routes added
      parameters:
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/VerifyNewAgentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Verify a new agent
      tags:
      - agent
      x-id: verify
...
  /auth/login:
    post:
      description: Initiates the login process by redirecting the user to the provider's
        login page
      produces:
      - text/plain
      responses:
        "302":
          description: Redirects to login page or IdP
          schema:
            type: string
        "403":
          description: 'Forbidden(webui): follow X-Redirect-To header'
          schema:
            type: string
        "429":
          description: Too Many Requests
          schema:
            type: string
      summary: Login
      tags:
      - auth
      x-id: login
...
  /cert/info:
    get:
      consumes:
      - application/json
      description: Get cert info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CertInfo'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get cert info
      tags:
      - cert
  /cert/renew:
    post:
      consumes:
      - application/json
      description: Renew cert
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Renew cert
      tags:
      - cert
...
  /favicon:
    get:
      consumes:
      - application/json
      description: Get favicon
      parameters:
      - description: URL of the route
        in: query
        name: url
        type: string
      - description: Alias of the route
        in: query
        name: alias
        type: string
      produces:
      - image/svg+xml
      - image/x-icon
      - image/png
      - image/webp
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/homepage.FetchResult'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get favicon
      tags:
      - v1
      x-id: favicon
  /file/content:
    get:
      consumes:
      - application/json
      description: Get file content
      parameters:
      - format: filename
        in: query
        name: filename
        required: true
        type: string
      - enum:
        - config
        - provider
        - middleware
        in: query
        name: type
        required: true
        type: string
        x-enum-comments:
          FileTypeConfig: '@name FileTypeConfig'
          FileTypeMiddleware: '@name FileTypeMiddleware'
          FileTypeProvider: '@name FileTypeProvider'
        x-enum-varnames:
        - FileTypeConfig
        - FileTypeProvider
        - FileTypeMiddleware
      produces:
      - application/json
      - application/godoxy+yaml
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get file content
      tags:
      - file
      x-id: get
    put:
      consumes:
      - application/json
      description: Set file content
      parameters:
      - description: Type
        enum:
        - config
        - provider
        - middleware
        in: query
        name: type
        required: true
        type: string
      - description: Filename
        in: query
        name: filename
        required: true
        type: string
      - description: File
        in: body
        name: file
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Set file content
      tags:
      - file
      x-id: set
...
  /icons:
    get:
      consumes:
      - application/json
      description: List icons
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Keyword
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/homepage.IconMetaSearch'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: List icons
      tags:
      - v1
      x-id: icons
...
swagger: "2.0"